requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-markdownfield==0.11.0
fake-useragent==1.4.0
h11==0.14.0
idna==3.6
Markdown==3.5.1
multidict==6.0.4
outcome==1.3.0.post0
Pillow==10.1.0
PySocks==1.7.1
requests==2.31.0
selenium==4.16.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.3
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-markdownfield==0.11.0
fake-useragent==1.4.0
geckodriver-autoinstaller==0.1.0
h11==0.14.0
idna==3.6
Markdown==3.5.1
multidict==6.0.4
outcome==1.3.0.post0
Pillow==10.1.0
PySocks==1.7.1
requests==2.31.0
selenium==4.16.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.3
